TW-97: apply m3 design to empty group chat
