+++ lib/pages/chat/direct_chat_empty_view.dart
@@ -1,0 +1,52 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_gen/gen_l10n/l10n.dart';
+
+class DirectChatEmptyView extends StatelessWidget {
+  
+  const DirectChatEmptyView({
+    Key? key,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      padding: const EdgeInsets.all(16),
+      constraints: const BoxConstraints(
+        maxWidth: 236,
+      ),
+      decoration: BoxDecoration(
+        color: Colors.black.withOpacity(0.04),
+        borderRadius: BorderRadius.circular(16),
+      ),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.center,
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          Text(
+            L10n.of(context)!.noMessageHereYet,
+            style: const TextStyle(
+              fontWeight: FontWeight.w700,
+              fontSize: 17,
+              color: Colors.black,
+            ),
+          ),
+          const SizedBox(height: 8),
+          Text(
+            L10n.of(context)!.sendMessageGuide,
+            textAlign: TextAlign.center,
+            style: const TextStyle(
+              fontWeight: FontWeight.w400,
+              fontSize: 15,
+              color: Color(0xFF818C99),
+            ),
+          ),
+          const SizedBox(height: 8),
+          Text(
+            '🤗',
+            style: Theme.of(context).textTheme.titleLarge?.merge(const TextStyle(fontSize: 88)),
+          ),
+        ],
+      ),
+    );
+  }
+}
